+++ src/common.ts
@@ -1,2 +1,4 @@
+import type { StandardSchemaV1 } from "@standard-schema/spec";
+
 export const noOp = () => {};
-export const NO_ERRORS = Object.freeze([] as string[]);
+export const NO_ERRORS = Object.freeze([] as StandardSchemaV1.Issue[]);