+++ src/validate-std-schema.ts
@@ -1,50 +1,12 @@
 import type { StandardSchemaV1 } from "@standard-schema/spec";
-import { FieldPath } from "./FieldPath.ts";
-import type { Issue } from "./validate.ts";
 
-export async function getValidationIssues<T>(data: T, validators: ReadonlyArray<StandardSchemaV1<T>>) {
+export async function getValidationIssues<T>(data: T, validators: ReadonlyArray<StandardSchemaV1<T>>): Promise<StandardSchemaV1.Issue[]> {
     const results = await Promise.all(validators.map(async validator => await _getValidationIssues(data, validator)));
     return results.flatMap(issues => issues);
 }
 
-async function _getValidationIssues<T>(data: T, validator: StandardSchemaV1<T>) {
+async function _getValidationIssues<T>(data: T, validator: StandardSchemaV1<T>): Promise<StandardSchemaV1.Issue[]> {
     const result = await validator["~standard"].validate(data);
     if (!result.issues) return [];
-
-    const nativeIssues: Issue[] = [];
-    for (const issue of result.issues) {
-        if (!issue.path) {
-            console.warn("Ignoring validation issue with no path:", JSON.stringify(issue));
-            continue;
-        }
-        const path = convertPath(issue.path);
-
-        nativeIssues.push({ path, message: issue.message })
-    }
-    return nativeIssues;
+    return result.issues.filter(issue => issue.path);
 }
-
-function convertPath(path: StdPath) {
-    let fieldPath = FieldPath.create();
-    path.map(part => unwrapPathPart(part)).forEach(part => {
-        if (typeof part === "string" || typeof part === "number") {
-            fieldPath = fieldPath.withProperty(part);
-        }
-        else {
-            throw new Error("Symbols are not supported");
-        }
-    });
-    return fieldPath;
-}
-
-function unwrapPathPart(part: PropertyKey | StdPath[number]): string | number | symbol {
-    if (typeof part === "object" && Object.hasOwn(part, "key")) {
-        return part.key;
-    }
-    else if (typeof part === "string" || typeof part === "number" || typeof part === "symbol") {
-        return part;
-    }
-    throw new Error("Unsupported path part " + part);
-}
-
-type StdPath = NonNullable<StandardSchemaV1.Issue["path"]>;